Add traditional Python optimization tier (NumPy vectorized search)
Adds an intermediate optimization step between the Python for-loop baseline
and Numba/FAISS, demonstrating core "basic Python" techniques:
- NumPy vectorized cosine similarity via matrix multiply (no Python loop)
- float32 dtype for 2x memory reduction vs float64
- np.argpartition for O(n) partial sort vs O(n log n) full sort
- Pre-normalized embeddings variant to skip redundant norm computation

Also fixes numba_search.py to use float32 consistently and updates
benchmark_stage3.py with the full 4-tier progression.

+++ optimized/stage3_search/numba_search.py
@@ -48,14 +48,14 @@ def search_similar_numba(
         Tuple of (indices, scores) arrays, each of length top_k.
     """
     n = embeddings.shape[0]
-    scores = np.empty(n, dtype=np.float64)
+    scores = np.empty(n, dtype=np.float32)
 
     for i in prange(n):
         scores[i] = cosine_similarity_numba(query_emb, embeddings[i])
 
     # Partial argsort — get top_k indices
     indices = np.argsort(scores)[::-1][:top_k]
-    top_scores = np.empty(top_k, dtype=np.float64)
+    top_scores = np.empty(top_k, dtype=np.float32)
     for i in range(top_k):
         top_scores[i] = scores[indices[i]]
